WIP: flat_scope -> scopes
