+++ src/lib.rs
@@ -9,7 +9,6 @@ extern crate lazy_static;
 extern crate log;
 
 pub mod diagnostic;
-mod flat_scope;
 pub mod js_regex;
 pub mod linter;
 pub mod rules;
