+++ src/linter.rs
@@ -2,14 +2,10 @@
 use crate::diagnostic::LintDiagnostic;
 use crate::diagnostic::Location;
 use crate::rules::LintRule;
-use crate::scopes::Scope;
-use crate::scopes::ScopeVisitor;
+use crate::scopes::{analyze, Scope};
 use crate::swc_util::get_default_ts_config;
 use crate::swc_util::AstParser;
-use crate::{
-  flat_scope::{analyze, FlatScope},
-  swc_util::SwcDiagnosticBuffer,
-};
+use crate::swc_util::SwcDiagnosticBuffer;
 use std::collections::HashMap;
 use std::rc::Rc;
 use std::sync::Arc;
@@ -22,7 +18,6 @@ use swc_common::BytePos;
 use swc_common::SourceMap;
 use swc_common::Span;
 use swc_ecmascript::parser::Syntax;
-use swc_ecmascript::visit::Visit;
 
 lazy_static! {
   static ref IGNORE_COMMENT_CODE_RE: regex::Regex =
@@ -37,9 +32,8 @@ pub struct Context {
   pub leading_comments: HashMap<BytePos, Vec<Comment>>,
   pub trailing_comments: HashMap<BytePos, Vec<Comment>>,
   pub ignore_directives: Vec<IgnoreDirective>,
-  pub root_scope: Scope,
   /// Arc as it's not modified
-  pub scope: Arc<FlatScope>,
+  pub scope: Arc<Scope>,
 }
 
 impl Context {
@@ -349,9 +343,6 @@ impl Linter {
       &leading,
     );
 
-    let mut scope_visitor = ScopeVisitor::default();
-    let root_scope = scope_visitor.get_root_scope();
-    scope_visitor.visit_module(&module, &module);
     let scope = Arc::new(analyze(&module));
 
     let context = Arc::new(Context {
@@ -361,7 +352,6 @@ impl Linter {
       leading_comments: leading,
       trailing_comments: trailing,
       ignore_directives,
-      root_scope,
       scope,
     });
