+++ src/rules/no_class_assign.rs
@@ -1,7 +1,7 @@
 // Copyright 2020 the Deno authors. All rights reserved. MIT license.
 use super::Context;
 use super::LintRule;
-use crate::{flat_scope::BindingKind, swc_util::find_lhs_ids};
+use crate::{scopes::BindingKind, swc_util::find_lhs_ids};
 use swc_ecmascript::ast::AssignExpr;
 use swc_ecmascript::visit::noop_visit_type;
 use swc_ecmascript::visit::Node;
