+++ src/rules/no_const_assign.rs
@@ -1,7 +1,7 @@
 // Copyright 2020 the Deno authors. All rights reserved. MIT license.
 use super::Context;
 use super::LintRule;
-use crate::flat_scope::BindingKind;
+use crate::scopes::BindingKind;
 use swc_common::Span;
 use swc_ecmascript::ast::AssignExpr;
 use swc_ecmascript::ast::Expr;
